pyproject.toml
[project]
name = "debt-collector-ai"
version = "0.1.0"
description = "Debt Collection agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepgram-sdk>=4.3.1",
    "livekit-agents>=1.1.1",
    "livekit-plugins-cartesia~=1.0",
    "livekit-plugins-deepgram~=1.0",
    "livekit-plugins-openai~=1.0",
    "twilio>=9.6.3",
]
